refactor: split coordinator.py into mixin files
Decompose the 3083-line TaskMateCoordinator class into 7 focused
mixin files (chores, assignments, rewards, points, timed, calendar,
templates). The main coordinator.py is now a slim 186-line
orchestrator that inherits from all mixins plus DataUpdateCoordinator.

No logic changes — pure structural move with per-file imports.
